requirements.txt
astroid==2.2.5
Django==2.2.2
django-bootstrap4==0.0.8
django-ckeditor==5.7.1
django-comments-xtd==2.4.0
django-contrib-comments==1.9.1
django-js-asset==1.2.2
django-materialize==1.0.1a2
djangorestframework==3.9.4
docutils==0.14
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
